+++ httprunner/__init__.py
@@ -1,16 +1,13 @@
-__version__ = "3.0.7"
+__version__ = "3.0.8"
 __description__ = "One-stop solution for HTTP(S) testing."
 
 from httprunner.runner import HttpRunner
-from httprunner.schema import TConfig, TStep
 from httprunner.testcase import Config, Step, RunRequest, RunTestCase
 
 __all__ = [
     "__version__",
     "__description__",
     "HttpRunner",
-    "TConfig",
-    "TStep",
     "Config",
     "Step",
     "RunRequest",