+++ httprunner/cli.py
@@ -4,12 +4,16 @@ import sys
 
 import pytest
 from loguru import logger
+from sentry_sdk import capture_message
 
 from httprunner import __description__, __version__
 from httprunner.compat import ensure_cli_args
 from httprunner.ext.har2case import init_har2case_parser, main_har2case
 from httprunner.make import init_make_parser, main_make
 from httprunner.scaffold import init_parser_scaffold, main_scaffold
+from httprunner.utils import init_sentry_sdk
+
+init_sentry_sdk()
 
 
 def init_parser_run(subparsers):
@@ -20,6 +24,7 @@ def init_parser_run(subparsers):
 
 
 def main_run(extra_args):
+    capture_message("start to run")
     # keep compatibility with v2
     extra_args = ensure_cli_args(extra_args)