+++ httprunner/client.py
@@ -11,6 +11,7 @@ from requests.exceptions import (
     MissingSchema,
     RequestException,
 )
+from sentry_sdk import capture_exception
 
 from httprunner.schema import RequestData, ResponseData
 from httprunner.schema import SessionData, ReqRespData
@@ -42,20 +43,21 @@ def get_req_resp_record(resp_obj: Response) -> ReqRespData:
     # record actual request info
     request_headers = dict(resp_obj.request.headers)
     request_cookies = resp_obj.request._cookies.get_dict()
-    request_body = resp_obj.request.body
-    try:
-        request_body = json.loads(request_body)
-    except json.JSONDecodeError:
-        # str: Unexpected UTF-8 BOM (decode using utf-8-sig)
-        pass
-    except UnicodeDecodeError:
-        # bytes/bytearray: request body in protobuf
-        pass
-    except TypeError:
-        # neither str nor bytes/bytearray, e.g. None
-        pass
 
-    if request_body:
+    request_body = resp_obj.request.body
+    if request_body is not None:
+        try:
+            request_body = json.loads(request_body)
+        except json.JSONDecodeError:
+            # str: a=1&b=2
+            pass
+        except UnicodeDecodeError as ex:
+            # bytes/bytearray: request body in protobuf
+            capture_exception(ex)
+        except TypeError as ex:
+            # neither str nor bytes/bytearray, e.g. <MultipartEncoder>
+            capture_exception(ex)
+
         request_content_type = lower_dict_keys(request_headers).get("content-type")
         if request_content_type and "multipart/form-data" in request_content_type:
             # upload file type