+++ httprunner/runner.py
@@ -46,6 +46,7 @@ class HttpRunner(object):
     __step_datas: List[StepData] = None
     __session: HttpSession = None
     __session_variables: VariablesMapping = {}
+    __export_variables: VariablesMapping = {}
     # time
     __start_at: float = 0
     __duration: float = 0
@@ -101,6 +102,7 @@ class HttpRunner(object):
         method = parsed_request_dict.pop("method")
         url_path = parsed_request_dict.pop("url")
         url = build_url(self.__config.base_url, url_path)
+        parsed_request_dict["verify"] = self.__config.verify
         parsed_request_dict["json"] = parsed_request_dict.pop("req_json", {})
 
         # request
@@ -147,6 +149,8 @@ class HttpRunner(object):
         except ValidationFailure:
             self.__session.data.success = False
             log_req_resp_details()
+            # log testcase duration before raise ValidationFailure
+            self.__duration = time.time() - self.__start_at
             raise
         finally:
             # save request & response meta data
@@ -248,6 +252,7 @@ class HttpRunner(object):
         self.__step_datas: List[StepData] = []
         self.__session = self.__session or HttpSession()
         self.__session_variables = {}
+        self.__export_variables = {}
 
         # run teststeps
         for step in self.__teststeps:
@@ -269,6 +274,11 @@ class HttpRunner(object):
             self.__session_variables.update(extract_mapping)
 
         self.__duration = time.time() - self.__start_at
+
+        self.__export_variables = self.get_export_variables()
+        if self.__export_variables:
+            logger.info(f"export variables: {self.__export_variables}")
+
         return self
 
     def run_path(self, path: Text) -> "HttpRunner":
@@ -293,6 +303,9 @@ class HttpRunner(object):
         return self.__step_datas
 
     def get_export_variables(self) -> Dict:
+        if self.__export_variables:
+            return self.__export_variables
+
         export_vars_mapping = {}
         for var_name in self.__config.export:
             if var_name not in self.__session_variables: